Transfert of scripts from bitbucket and new functionnalities
